requirements.txt
PyQt6>=6.4.0,<7.0.0
google-api-python-client>=2.0.0,<3.0.0
google-auth-oauthlib>=0.5.0,<2.0.0
google-auth-httplib2>=0.1.0,<0.2.0
plyer>=2.1.0,<3.0.0
beautifulsoup4>=4.10.0,<5.0.0
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0
pandas>=1.4.0,<3.0.0
openpyxl>=3.0.0,<4.0.0 # Required by pandas for xlsx
requests>=2.20.0,<3.0.0 # Often a dependency of google libs, good to specify